visualization/requirements.txt
absl-py==2.1.0
accelerate @ git+https://github.com/huggingface/accelerate@724824abbe0aed8606661bbce5e057c0d2447794
addict==2.4.0
aiofiles==23.2.1
aiohttp==3.9.3
aiosignal==1.3.1
albumentations==1.4.4
altair==5.2.0
aniposelib @ file:///comp_robot/zhangyuhong1/code/aistplusplus_smplx/aniposelib-master.zip#sha256=9394fe89473c74f73a7bfdd6dbc4f213336fd7597d48a448bd94490ec06c7276
annotated-types==0.6.0
antlr4-python3-runtime==4.9.3
anyio==4.3.0
asttokens==2.4.1
async-timeout==4.0.3
attrs==23.2.0
audioread==3.0.1
av==11.0.0
beartype==0.18.5
beautifulsoup4==4.12.3
black==24.2.0
blinker==1.8.2
braceexpand==0.1.7
cachetools @ file:///home/conda/feedstock_root/build_artifacts/cachetools_1708987703938/work
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
chumpy @ git+https://github.com/mattloper/chumpy@51d5afd92a8ded3637553be8cef41f328a1c863a
click==8.1.7
cloudpickle==3.0.0
colorama==0.4.6
colored==2.2.4
coloredlogs==15.0.1
comm==0.2.2
ConfigArgParse==1.7
contourpy==1.2.0
coverage==7.4.4
cycler==0.12.1
Cython==3.0.9
dash==2.17.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
decorator==5.1.1
detectron2 @ git+https://github.com/facebookresearch/detectron2.git@3ff5dd1cff4417af07097064813c9f28d7461d3c
dill==0.3.8
distro==1.9.0
docker-pycreds==0.4.0
einops==0.7.0
exceptiongroup==1.2.0
executing==2.0.1
fastapi==0.110.0
fastjsonschema==2.19.1
ffmpeg-python==0.2.0
ffmpy==0.3.2
filelock==3.13.1
fire==0.1.3
flake8==7.0.0
Flask==3.0.3
flatbuffers==24.3.7
fonttools==4.49.0
freetype-py==2.4.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
fvcore==0.1.5.post20221221
gitdb==4.0.11
GitPython==3.1.43
gradio==4.21.0
gradio_client==0.12.0
grpcio==1.62.0
h11==0.14.0
httpcore==1.0.4
httpx==0.27.0
huggingface-hub==0.21.2
humanfriendly==10.0
humanize==4.9.0
hydra-core==1.3.2
idna==3.6
imageio==2.34.0
imageio-ffmpeg==0.4.9
importlib_metadata==7.1.0
importlib_resources==6.3.0
iniconfig==2.0.0
interrogate==1.7.0
iopath==0.1.9
ipython==8.25.0
ipywidgets==8.1.3
isort==4.3.21
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.3
joblib==1.3.2
json-tricks==3.17.3
jsonlines==4.0.0
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
kiwisolver==1.4.5
lazy_loader==0.3
librosa==0.9.2
lightning-utilities==0.10.1
llvmlite==0.42.0
loguru==0.7.2
lpips==0.1.4
Markdown==3.5.2
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mccabe==0.7.0
mdurl==0.1.2
mpmath==1.3.0
multidict==6.0.5
multiprocess==0.70.16
munkres==1.1.4
mypy-extensions==1.0.0
nbformat==5.10.4
nest-asyncio==1.6.0
networkx==3.2.1
neural_renderer_pytorch @ git+https://github.com/shubham-goel/NMR.git@e990b3c70f48d39231f607c79d76ce3db4bf7483
numba==0.59.1
numpy==1.26.4
nvidia-ml-py @ file:///home/conda/feedstock_root/build_artifacts/nvidia-ml-py_1698947663801/work
nvitop @ file:///home/conda/feedstock_root/build_artifacts/nvitop_1702822107828/work
omegaconf==2.3.0
onnxruntime==1.17.1
open3d==0.18.0
openai==1.25.2
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
opendr @ git+https://gitlab.eecs.umich.edu/ngv-python-modules/opendr.git@dc66c3e9f1e8a8d4df3acabfceb2a6b7a09b937c
orjson==3.9.15
p_tqdm==1.4.0
packaging==23.2
pandas==2.2.1
parameterized==0.9.0
parso==0.8.4
pathos==0.3.2
pathspec==0.12.1
pexpect==4.9.0
-e git+https://github.com/brjathu/PHALP.git@6a0ef228714093dc6278df101d97884352210e68#egg=phalp
pillow==10.2.0
platformdirs==4.2.0
plotly==5.22.0
pluggy==1.4.0
pooch==1.8.1
portalocker==2.8.2
pox==0.3.4
ppft==1.7.6.8
prompt_toolkit==3.0.45
protobuf==4.25.3
psutil @ file:///home/conda/feedstock_root/build_artifacts/psutil_1705722392846/work
ptyprocess==0.7.0
pure-eval==0.2.2
py==1.11.0
pycocotools==2.0.7
pycodestyle==2.11.1
pycparser==2.22
pydantic==2.6.4
pydantic_core==2.16.3
pydub==0.25.1
pyflakes==3.2.0
pyglet==2.0.12
Pygments==2.17.2
PyOpenGL @ git+https://github.com/mmatl/pyopengl.git@76d1261adee2d3fd99b418e75b0416bb7d2865e6
pyparsing==3.1.1
pyquaternion==0.9.9
pyrender==0.1.45
PySocks==1.7.1
pytest==8.1.1
pytest-runner==6.0.1
python-dateutil==2.8.2
python-multipart==0.0.9
pytorch-lightning==2.2.0.post0
pytorch3d==0.7.2
pytube @ git+https://github.com/pytube/pytube.git@a32fff39058a6f7e5e59ecd06a7467b71197ce35
pytz==2024.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
resampy==0.4.3
retrying==1.3.4
rich==13.7.1
rpds-py==0.18.0
ruff==0.3.3
safetensors==0.4.2
scenedetect==0.6.2
scipy==1.12.0
seaborn==0.13.2
semantic-version==2.10.0
sentry-sdk==2.1.1
setproctitle==1.3.3
shapely==2.0.4
shellingham==1.5.4
six==1.16.0
smmap==5.0.1
smplx==0.1.28
sniffio==1.3.1
SoundFile==0.10.3.post1
soupsieve==2.5
stack-data==0.6.3
starlette==0.36.3
sympy==1.12
tabulate==0.9.0
tenacity==8.3.0
termcolor @ file:///home/conda/feedstock_root/build_artifacts/termcolor_1704357939450/work
terminaltables==3.1.10
threadpoolctl==3.3.0
tifffile==2024.2.12
timm==0.9.16
toml==0.10.2
tomli==2.0.1
tomlkit==0.12.0
toolz==0.12.1
torch==1.13.1+cu116
torchgeometry==0.1.2
torchmetrics==1.3.1
torchvision==0.14.1+cu116
tqdm==4.45.0
traitlets==5.14.3
trimesh==4.1.6
typer==0.9.0
typing_extensions==4.10.0
tzdata==2024.1
Unidecode==1.1.1
urllib3==2.2.1
uvicorn==0.28.0
wandb==0.17.0
wcwidth==0.2.13
webdataset==0.2.86
websockets==11.0.3
Werkzeug==3.0.1
wget==3.2
widgetsnbextension==4.0.11
xdoctest==1.1.3
xtcocotools==1.14.3
yacs==0.1.8
yapf==0.40.2
yarl==1.9.4
zipp==3.18.1
